requirements.txt
fastapi
uvicorn
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
python-multipart
python-dotenv
